+++ udata_front/theme/gouvfr/assets/js/components/Reuse/Reuse.vue
@@ -3,7 +3,7 @@
 </template>
 
 <script setup lang="ts">
-import  { ReuseCard, type Reuse } from '@datagouv/components';
+import  { ReuseCard, type Reuse } from '@datagouv/components/ts';
 
 defineProps<{
   reuse: Reuse,