+++ udata_front/theme/gouvfr/assets/js/components/Reuse/ReuseList.vue
@@ -25,7 +25,7 @@
 </template>
 
 <script lang="ts">
-import type { Reuse } from "@datagouv/components";
+import type { Reuse } from "@datagouv/components/ts";
 export type ReuseListProps = {
   reuses: Array<Reuse>,
   loading?: boolean,
@@ -39,8 +39,8 @@ const defaultPageSize = 20;
 
 <script setup lang="ts">
 import { ref } from 'vue';
-import { Pagination } from '@datagouv/components';
 import ReuseCard from "../Reuse/Reuse.vue";
+import { Pagination } from '@datagouv/components/ts';
 import Loader from "./loader.vue";
 
 withDefaults(defineProps<ReuseListProps>(), {