+++ udata_front/theme/gouvfr/assets/js/pages/OrganizationPublishingForm/Step2DescribeOrganization.vue
@@ -227,7 +227,7 @@
   </div>
 </template>
 <script lang="ts">
-  import type { NewOrganization } from '@datagouv/components';
+  import type { NewOrganization } from '@datagouv/components/ts';
   import type { OrganizationV1 } from '../../types';
 
   export type DescribeOrganizationProps = {
@@ -238,7 +238,7 @@
   };
 </script>
 <script setup lang="ts">
-  import { Well } from '@datagouv/components';
+  import { Well } from '@datagouv/components/ts';
   import { url } from '@vuelidate/validators';
   import axios from 'axios';
   import { computed, reactive, ref, watchEffect } from 'vue';