+++ udata_front/theme/gouvfr/datagouv-components/src/components/ResourceAccordion/ResourceAccordion.vue
@@ -1,391 +1,168 @@
 <template>
-  <article :class="{'drop-shadow': expanded}">
+  <div class="border border-default-grey" :class="{ 'fr-pb-4v': open }">
     <header
-      class="fr-p-5v fr-grid-row fr-grid-row--middle no-wrap wrap-md justify-between border-default-grey"
-      :class="{'border-bottom': !expanded}"
+      class="fr-p-4v flex items-center justify-between relative"
       :id="resourceHeaderId"
     >
-      <div class="fr-col-auto fr-grid-row fr-grid-row--top no-wrap">
-        <div class="fr-col-auto fr-mr-2w fr-icon-svg fr-icon--sm">
-          <img :src="url" alt="" />
+      <div>
+        <h4 class="fr-mb-1v flex items-center" :id="resourceTitleId">
+          <button type="button" @click="toggle" class="fr-p-0 flex items-center" data-testid="expand-button" :aria-expanded="open">
+            <ResourceIcon :resource class="fr-icon--xs fr-mr-1v" />
+            <span :class="{
+              'font-bold': open,
+            }"><TextClamp :max-lines="1" :text="resource.title || t('Nameless file')"/></span>
+
+            <span class="absolute inset-0 z-1"></span>
+          </button>
+          <CopyButton :label="$t('Copy link')" :copied-label="$t('Link copied!')" :text="resourceExternalUrl" class="z-2" />
+        </h4>
+        <div class="text-grey-380 subheaders-infos">
+          <SchemaBadge :resource class="dash-after" />
+          <span class="fr-text--xs fr-mb-0 dash-after">{{ t('Updated {date}', {date: formatRelativeIfRecentDate(lastUpdate)}) }}</span>
+          <span v-if="resource.format" class="fr-text--xs fr-mb-0 dash-after">
+            <span class="hidden show-on-small">{{ t("Format") }}</span>
+            {{ resource.format.trim().toLowerCase() }}
+            <span v-if="resource.filesize">({{ filesize(resource.filesize) }})</span>
+          </span>
+          <span class="inline-flex items-center fr-text--xs fr-mb-0" :aria-label="t('{n} downloads', resource.metrics.views)">
+            <span class="fr-icon-download-line fr-icon--xs fr-mr-1v"></span>
+            <span>{{ summarize(resource.metrics.views) }} <span class="hidden show-on-small">{{ t("downloads") }}</span></span>
+          </span>
         </div>
-        <div class="fr-col-auto">
-          <h4
-            class="fr-mb-1v"
-            :id="resourceTitleId"
-          >
-            {{ resource.title || t('Nameless file') }}
-          </h4>
-          <div class="fr-my-0 text-grey-380 fr-grid-row fr-grid-row--middle">
-            <p v-if="hasSchema" class="fr-tag fr-tag--sm z-2">
-              <strong class="fr-mr-1v">{{ t("schema:") }}</strong>{{schemaName ? schemaName : schemaUrl}}
-            </p>
-            <p
-              v-if="hasSchema && hasSchemaErrors"
-              class="fr-tag fr-tag--error fr-tag--on-right fr-tag--sm fr-ml-n3v fr-pl-2w fr-icon-error-line fr-tag--icon-left"
-            >
-              {{ t("Invalid") }}
-            </p>
-            <p :class="{'dash-before': hasSchema}" class="fr-text--xs fr-m-0 dash-after">{{ t('Updated {date}', {date: formatRelativeIfRecentDate(lastUpdate)}) }}</p>
-            <p v-if="resource.format" class="fr-text--xs fr-m-0 dash-after">
-              {{ resource.format?.trim()?.toLowerCase() }}
-              <template v-if="resource.filesize">({{ filesize(resource.filesize) }})</template>
-            </p>
-            <p class="fr-text--xs fr-m-0">{{ t('{n} downloads', resource.metrics.views || 0) }}</p>
-          </div>
-          <p class="fr-mb-0 fr-mt-1v fr-text--xs text-grey-380" v-if="checkIsCommunityResource(resource)">
-            {{ t('From') }}
-            <a class="fr-link" v-if="resource.organization" :href="resource.organization.page" >
-              <OrganizationNameWithCertificate :organization="resource.organization" />
-            </a>
-            <template v-else-if="owner">{{owner}}</template>
-          </p>
-        </div>
+        <p class="fr-mb-0 fr-mt-1v fr-text--xs text-grey-380" v-if="communityResource">
+          {{ t('From') }}
+          <a class="fr-link fr-text--xs" v-if="communityResource.organization" :href="communityResource.organization.page" >
+            <OrganizationNameWithCertificate :organization="communityResource.organization" />
+          </a>
+          <template v-else-if="owner">{{owner}}</template>
+        </p>
       </div>
-      <div class="fr-col-auto fr-ml-auto">
-        <div class="fr-grid-row fr-grid-row--middle no-wrap wrap-md">
-          <p class="text-default-warning fr-m-0" v-if="unavailable">
-            {{t('Unavailable')}}
-          </p>
-          <p class="fr-col-auto fr-ml-3v fr-m-0">
-            <button
-              @click="expand"
-              role="button"
-              :aria-expanded="expanded"
-              :aria-controls="resourceContentId"
-              class="fr-btn fr-btn--tertiary-no-outline fr-btn--icon-left fr-btn--secondary-grey-500"
-              :class="{'fr-icon-arrow-up-s-line': expanded, 'fr-icon-arrow-down-s-line': !expanded}"
-              data-testid="expand-button"
-            >
-              <template v-if="expanded">
-                {{ t('Close details') }}
-              </template>
-              <template v-else>
-                {{ hasPreview ? t('See data') : t('See metadata') }}
-              </template>
-            </button>
-          </p>
-          <p class="fr-col-auto fr-ml-3v fr-m-0" v-if="resource.format === 'url'">
-            <a
-              :href="resource.latest"
-              :title="t('File link - opens a new window')"
-              rel="ugc nofollow noopener"
-              target="_blank"
-              class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-icon-external-link-line fr-icon--sm"
-            >
-            </a>
-          </p>
-          <p class="fr-col-auto fr-ml-3v fr-m-0" v-else>
-            <a
-              :href="resource.latest"
-              rel="ugc nofollow noopener"
-              :title="t('Download file')"
-              download
-              class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-icon-download-line fr-icon--sm matomo_download"
-            >
-            </a>
-          </p>
-          <p class="fr-col-auto fr-ml-3v fr-m-0" v-if="canEdit">
-            <EditButton
-              :dataset-id="datasetId"
-              :resource-id="resource.id"
-              :is-community-resource="isCommunityResource"
-            />
-          </p>
-        </div>
+      <div class="flex items-center fr-ml-4v buttons">
+        <p class="text-default-warning fr-m-0 fr-mr-2v" v-if="unavailable">
+          {{ t('Unavailable') }}
+        </p>
+        <p class="fr-col-auto fr-ml-3v fr-m-0 z-2" v-if="resource.format === 'url'">
+          <a
+            :href="resource.latest"
+            :title="t('File link - opens a new window')"
+            rel="ugc nofollow noopener"
+            target="_blank"
+            class="fr-btn fr-btn--sm"
+          >
+            {{ $t('Visit') }}
+          </a>
+        </p>
+        <p class="fr-col-auto fr-ml-3v fr-m-0" v-else>
+          <a
+            :href="resource.latest"
+            rel="ugc nofollow noopener"
+            :title="t('Download file')"
+            download
+            class="relative text-transform-uppercase fr-btn fr-btn--sm fr-btn--icon-left fr-icon-test-tube-line fr-icon-download-line fr-icon--sm matomo_download z-2"
+          >
+            {{ format }}
+          </a>
+        </p>
+        <p class="fr-col-auto fr-ml-3v fr-m-0 z-2" v-if="canEdit">
+          <EditButton
+            :dataset-id="datasetId"
+            :resource-id="resource.id"
+            :is-community-resource="isCommunityResource"
+          />
+        </p>
+        <div
+          class="fr-icon--sm fr-ml-4v"
+          :class="{'fr-icon-arrow-up-s-line': open, 'fr-icon-arrow-down-s-line': !open}"
+        ></div>
       </div>
     </header>
     <section
-      class="accordion-content"
+      v-if="open"
       :aria-labelledby="resourceTitleId"
       :id="resourceContentId"
-      ref="contentRef"
     >
-      <div class="fr-tabs fr-tabs--no-border fr-mt-1v" ref="tabsRef">
-        <ul class="fr-tabs__list" ref="tabListRef" role="tablist" :aria-label="t('File menu')">
-          <template v-if="hasPreview">
-            <li role="presentation">
-              <button
-                :id="getIdFromIndex(resourcePreviewIndex)"
-                class="fr-tabs__tab"
-                :tabindex="getTabPanelTabIndex(resourcePreviewIndex)"
-                role="tab"
-                :aria-selected="isSelected(resourcePreviewIndex)"
-                :aria-controls="getTabIdFromIndex(resourcePreviewIndex)"
-                @click="selectTab(resourcePreviewIndex)"
-              >
-                {{t('Preview')}}
-              </button>
-            </li>
-          </template>
-          <template v-if="hasPreview || hasSchema">
-            <li role="presentation">
-              <button
-                :id="getIdFromIndex(resourceStructureIndex)"
-                class="fr-tabs__tab"
-                :tabindex="getTabPanelTabIndex(resourceStructureIndex)"
-                role="tab"
-                :aria-selected="isSelected(resourceStructureIndex)"
-                :aria-controls="getTabIdFromIndex(resourceStructureIndex)"
-                @click="selectTab(resourceStructureIndex)"
-              >
-                {{t('Data structure')}}
-              </button>
-            </li>
-          </template>
-          <li role="presentation">
-            <button
-              :id="getIdFromIndex(resourceInformationIndex)"
-              class="fr-tabs__tab"
-              :tabindex="getTabPanelTabIndex(resourceInformationIndex)"
-              role="tab"
-              :aria-selected="isSelected(resourceInformationIndex)"
-              :aria-controls="getTabIdFromIndex(resourceInformationIndex)"
-              @click="selectTab(resourceInformationIndex)"
-            >
-              {{t('Metadata')}}
-            </button>
-          </li>
-        </ul>
-        <transition
-          name="slide-fade"
-          mode="in-out"
-          v-if="hasPreview"
-        >
-          <div
-            :id="getTabIdFromIndex(resourcePreviewIndex)"
-            class="fr-tabs__panel fr-tabs__panel--no-padding fr-table--dense"
-            :class="getTabPanelClass(resourcePreviewIndex)"
-            role="tabpanel"
-            :aria-labelledby="getIdFromIndex(resourcePreviewIndex)"
-            :tabindex="getTabPanelTabIndex(resourcePreviewIndex)"
-            v-show="isSelected(resourcePreviewIndex)"
-          >
-            <Preview  v-if="expanded" :resource="resource" />
-          </div>
-        </transition>
-        <transition
-          name="slide-fade"
-          mode="in-out"
-          v-if="hasPreview || hasSchema"
-        >
-          <div
-            :id="getTabIdFromIndex(resourceStructureIndex)"
-            class="fr-tabs__panel"
-            :class="getTabPanelClass(resourceStructureIndex)"
-            role="tabpanel"
-            :aria-labelledby="getIdFromIndex(resourceStructureIndex)"
-            :tabindex="getTabPanelTabIndex(resourceStructureIndex)"
-            v-show="isSelected(resourceStructureIndex)"
-          >
-            <DataStructure v-if="hasPreview" :resource="resource" />
-            <hr class="fr-my-5v fr-p-1v" v-if="hasPreview && hasSchema"/>
-            <template v-if="hasSchema">
-              <div class="fr-grid-row fr-grid-row--gutters fr-grid-row--middle">
-                <div class="fr-col-12 fr-col-lg">
-                  <h5 class="fr-h6 fr-mb-5v">{{t('Data schema')}}</h5>
-                  <ul class="fr-tags-group" v-if="hasSchemaErrors">
-                    <li v-for="[_key, schemaError] in schemaReport">
-                      <p class="fr-tag fr-tag--sm">{{schemaError.name}}</p>
-                    </li>
-                  </ul>
-                  <template v-else>
-                    <div class="fr-grid-row fr-grid-row--middle fr-mb-1w">
-                      <p class="fr-text--xs fr-m-0">{{t('This file is following a schema: ')}}</p>
-                      <p class="fr-tag fr-tag--sm fr-icon-checkbox-circle-line fr-tag--icon-left">
-                        {{schemaName ? schemaName : schemaUrl}}
-                      </p>
-                    </div>
-                      <i18n-t keypath="Schemas allow to describe data models, discover how schemas improve your data quality and the available use cases on {address}" class="fr-text--xs fr-m-0" tag="p" scope="global">
-                        <template #address>
-                          <a :href="config.schema_documentation_url">schema.data.gouv.fr</a>
+      <TabGroup size="sm" @change="switchTab">
+        <div class="fr-pl-4v fr-pr-4v fr-pb-4v" >
+          <TabList style="max-width: 100%; overflow-y: auto;">
+            <Tab v-for="tab in tabsOptions" :key="tab.key">{{ tab.label }}</Tab>
+          </TabList>
+        </div>
+        <TabPanels>
+          <TabPanel v-for="tab in tabsOptions" :key="tab.key">
+            <div v-if="tab.key === 'data'">
+              <Preview :resource="resource" />
+            </div>
+            <div v-if="tab.key === 'description'"  class="fr-pl-4v fr-pr-4v">
+              <div class="fr-mt-0 markdown fr-text--sm text-mention-grey" v-html="markdown(resource.description || '')"></div>
+            </div>
+            <div v-if="tab.key === 'data-structure'" class="fr-pl-4v fr-pr-4v">
+              <DataStructure v-if="hasPreview" :resource="resource" />
+            </div>
+            <div v-if="tab.key === 'metadata'" class="fr-pl-4v fr-pr-4v">
+              <Metadata :resource />
+            </div>
+            <div v-if="tab.key === 'downloads'" class="fr-pl-4v fr-pr-4v">
+              <dl class="fr-pl-0">
+                <dt class="font-bold fr-text--sm fr-mb-0" v-if="resource.format === 'url'">{{ $t('Original URL') }}</dt>
+                <dt class="font-bold fr-text--sm fr-mb-0" v-else>{{ $t('Original format') }}</dt>
+                <dd class="fr-text--sm fr-ml-0 fr-mt-0 fr-mb-2w text-mention-grey h-4w fr-grid-row fr-grid-row--middle">
+                  <span v-if="resource.format === 'url'">
+                    <a :href="resource.latest" class="fr-link fr-link--no-after" rel="ugc nofollow noopener" target="_blank">
+                      <TextClamp :auto-resize="true" :max-lines="1" :text="resource.url">
+                        <template #after>
+                          <span class="fr-ml-1v fr-icon-external-link-line fr-icon--sm"></span>
                         </template>
-                      </i18n-t>
-                  </template>
-                </div>
-                <div class="fr-col-auto">
-                  <div v-if="loading">
-                    <SchemaLoader/>
-                  </div>
-                  <div v-else>
-                    <a
-                      v-if="authorizeValidation"
-                      class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-mr-3v fr-btn--icon-left fr-icon-checkbox-line"
-                      :href="validationUrl"
-                    >
-                      {{ t('See validation report') }}
+                      </TextClamp>
                     </a>
-                    <a
-                      v-if="schemaName"
-                      class="fr-btn fr-btn--secondary fr-btn--secondary-grey-500 fr-btn--icon-left fr-icon-book-2-line"
-                      :href="documentationUrl"
-                    >
-                      {{ t('See schema documentation') }}
+                  </span>
+                  <span v-else>
+                    <span class="text-blue-400 fr-icon-download-line fr-icon--sm fr-mr-1v fr-mt-1v"></span>
+                    <a :href="resource.latest" class="fr-link" rel="ugc nofollow noopener">
+                      <span>{{ $t('Format {format}', { format: resource.format }) }}<span v-if="resource.filesize"> - {{ filesize(resource.filesize) }}</span></span>
                     </a>
-                  </div>
-                </div>
-              </div>
-            </template>
-          </div>
-        </transition>
-        <transition
-          name="slide-fade"
-          mode="in-out"
-        >
-          <div
-            :id="getTabIdFromIndex(resourceInformationIndex)"
-            class="fr-tabs__panel"
-            :class="getTabPanelClass(resourceInformationIndex)"
-            role="tabpanel"
-            :aria-labelledby="getIdFromIndex(resourceInformationIndex)"
-            :tabindex="getTabPanelTabIndex(resourceInformationIndex)"
-            v-show="isSelected(resourceInformationIndex)"
-          >
-            <div class="fr-grid-row fr-grid-row--gutters">
-              <DescriptionList>
-                <DescriptionTerm>{{ t('URL') }}</DescriptionTerm>
-                <DescriptionDetails :withEllipsis="false">
-                  <div class="fr-col text-overflow-ellipsis">
-                    <a :href="resource.url">{{resource.url}}</a>
-                  </div>
-                  <div class="fr-ml-1w fr-col-auto">
-                    <CopyButton :text="resource.url"/>
-                  </div>
-                </DescriptionDetails>
-                <DescriptionTerm>{{ t('Permalink') }}</DescriptionTerm>
-                <DescriptionDetails :withEllipsis="false">
-                  <div class="fr-col text-overflow-ellipsis">
-                    <a :href="resource.latest">{{resource.latest}}</a>
-                  </div>
-                  <div class="fr-ml-1w fr-col-auto">
-                    <CopyButton :text="resource.latest"/>
-                  </div>
-                </DescriptionDetails>
-                <DescriptionTerm>{{ t('ID') }}</DescriptionTerm>
-                <DescriptionDetails :withEllipsis="false">
-                  <div class="fr-col text-overflow-ellipsis">
-                    {{ resource.id }}
-                  </div>
-                  <div class="fr-ml-1w fr-col-auto">
-                    <CopyButton :text="resource.id"/>
-                  </div>
-                </DescriptionDetails>
-                <template v-if="resource.mime">
-                  <DescriptionTerm>{{ t('MIME Type') }}</DescriptionTerm>
-                  <DescriptionDetails>
-                    {{resource.mime}}
-                  </DescriptionDetails>
-                </template>
-              </DescriptionList>
-              <DescriptionList>
-                <DescriptionTerm>{{ t('Created on') }}</DescriptionTerm>
-                <DescriptionDetails>
-                  {{formatDate(resource.created_at)}}
-                </DescriptionDetails>
-                <DescriptionTerm>{{ t('Modified on') }}</DescriptionTerm>
-                <DescriptionDetails>
-                  {{formatDate(resource.last_modified)}}
-                </DescriptionDetails>
-                <template v-if="resource.checksum">
-                  <DescriptionTerm>{{resource.checksum.type}}</DescriptionTerm>
-                  <DescriptionDetails :withEllipsis="false">
-                    <div class="fr-col text-overflow-ellipsis">
-                      {{resource.checksum.value}}
-                    </div>
-                    <div class="fr-col-auto">
-                      <CopyButton :text="resource.checksum.value"/>
-                    </div>
-                  </DescriptionDetails>
-                </template>
-              </DescriptionList>
-              <DescriptionList>
-                <template v-if="resource.filesize">
-                  <DescriptionTerm>{{ t('Size') }}</DescriptionTerm>
-                  <DescriptionDetails>
-                    {{ filesize(resource.filesize) }}
-                  </DescriptionDetails>
-                </template>
-              </DescriptionList>
-              <div class="fr-col-auto fr-ml-auto" v-if="config.show_copy_resource_permalink">
-                <button
-                  :id="resourceCopyId"
-                  ref="copyRef"
-                  :data-clipboard-text="resourceExternalUrl"
-                  class="fr-btn fr-btn--sm fr-btn--secondary fr-btn--secondary-grey-500 fr-btn--icon-right fr-icon-links-fill"
-                >
-                  {{t('Copy resource permalink')}}
-                </button>
-              </div>
+                  </span>
+                  <CopyButton :label="$t('Copy link')" :copied-label="$t('Link copied!')" :text="resource.latest" class="relative" />
+                </dd>
+              </dl>
             </div>
-            <template v-if="resource.description">
-              <h5 class="fr-text--sm fr-my-0 fr-text--bold">
-                {{ t("File description") }}
-              </h5>
-              <div class="fr-mt-0 markdown fr-text--sm text-mention-grey" v-html="markdown(resource.description)">
-              </div>
-            </template>
-            <div
-              class="text-align-right"
-              v-if="!hasPreview && resource.preview_url"
-            >
-              <a
-                :href="resource.preview_url"
-                class="fr-btn fr-btn--icon-left fr-icon-test-tube-line"
-              >
-                {{ t("Explore data") }}
-              </a>
-            </div>
-            <ExtraAccordion
-              class="fr-pt-3w fr-mt-3w border-top border-default-grey"
-              :button-text="t('See extras')"
-              :title-text="t('Resource Extras')"
-              title-level="h5"
-              :extra="resource.extras"
-              v-if="hasExtras"
-            />
-          </div>
-        </transition>
-      </div>
+          </TabPanel>
+        </TabPanels>
+      </TabGroup>
     </section>
-  </article>
+  </div>
 </template>
 
 <script setup lang="ts">
-import Clipboard from "clipboard";
-import { ref, computed, unref, onMounted, type ComputedRef } from "vue";
+import { ref, computed } from "vue";
 import { useI18n } from "vue-i18n";
 import DataStructure from "./DataStructure/DataStructure.vue";
 import Preview from "./Preview/Preview.vue";
-import SchemaLoader from "./SchemaLoader.vue";
-import ExtraAccordion from "../ExtraAccordion/ExtraAccordion.vue";
-import CopyButton from "../CopyButton/CopyButton.vue";
 import EditButton from "./EditButton.vue";
-import DescriptionDetails from "../DescriptionList/DescriptionDetails.vue";
-import DescriptionList from "../DescriptionList/DescriptionList.vue";
-import DescriptionTerm from "../DescriptionList/DescriptionTerm.vue";
 import OrganizationNameWithCertificate from "../Organization/OrganizationNameWithCertificate.vue";
 import useOwnerName from "../../composables/organizations/useOwnerName";
-import useResourceImage from "../../composables/resources/useResourceImage";
-import useSchema from "../../composables/resources/useSchema";
-import useTabs from "../../composables/useTabs";
 import { config } from "../../config";
-import { filesize, formatRelativeIfRecentDate, formatDate, markdown } from "../../helpers";
+import { filesize, formatRelativeIfRecentDate, markdown, summarize } from "../../helpers";
 import type { CommunityResource, Resource } from "../../types/resources";
-import useAccordion from "../../composables/useAccordion";
+import ResourceIcon from "./ResourceIcon.vue";
+import SchemaBadge from "./SchemaBadge.vue";
+import Metadata from "./Metadata.vue";
+import TextClamp from 'vue3-text-clamp';
+import TabGroup from "../Tabs/TabGroup.vue";
+import TabList from "../Tabs/TabList.vue";
+import Tab from "../Tabs/Tab.vue";
+import TabPanels from "../Tabs/TabPanels.vue";
+import TabPanel from "../Tabs/TabPanel.vue";
+import { trackEvent } from "../../helpers/matomo";
+import CopyButton from "../CopyButton/CopyButton.vue";
+import { getResourceFormatIconSvg } from "../../helpers/resources";
 
-type Props = {
+const props = withDefaults(defineProps<{
   datasetId: string,
   expandedOnMount?: boolean,
   isCommunityResource?: boolean,
   resource: Resource | CommunityResource,
   canEdit?: boolean
-};
-
-defineOptions({
-  inheritAttrs: false,
-});
-
-const props = withDefaults(defineProps<Props>(), {
+}>(), {
   expandedOnMount: false,
   isCommunityResource: false,
   canEdit: false,
@@ -393,116 +170,128 @@ const props = withDefaults(defineProps<Props>(), {
 
 const { t } = useI18n();
 
-function checkIsCommunityResource(_resource: Resource | CommunityResource): _resource is CommunityResource {
-  return props.isCommunityResource;
-}
+const hasPreview = computed(() => {
+  return config.tabular_api_url
+    && props.resource.extras['analysis:parsing:finished_at']
+    && !props.resource.extras['analysis:parsing:error']
+    && (config.tabular_allow_remote || props.resource.filetype === "file");
+})
 
-let owner: ReturnType<typeof useOwnerName>;
-if(checkIsCommunityResource(props.resource)) {
-  owner = useOwnerName(props.resource);
-}
-const { url } = useResourceImage(props.resource);
+const format = computed(() => getResourceFormatIconSvg(props.resource) ? props.resource.format : t("File"))
 
-const hasPreview: boolean = config.tabular_api_url &&
-  props.resource.extras['analysis:parsing:finished_at'] &&
-  !props.resource.extras['analysis:parsing:error'] &&
-  (config.tabular_allow_remote || props.resource.filetype === "file");
+const open = ref(props.expandedOnMount);
+const toggle = () => {
+  open.value = ! open.value;
 
-const contentRef = ref<HTMLElement | null>(null);
-const copyRef = ref<HTMLButtonElement | null>(null);
-const tabsRef = ref<HTMLButtonElement | null>(null);
-const tabListRef = ref<HTMLUListElement | null>(null);
-
-const schemaName = props.resource.schema ? props.resource.schema.name ?? "" : "";
-const schemaUrl = props.resource.schema ? props.resource.schema.url ?? "" : "";
-
-const hasExtras = Object.keys(props.resource.extras).length > 0;
-
-const hasSchema = schemaName || schemaUrl;
-const resourcePreviewIndex = 0;
-const resourceStructureIndex = hasPreview ? 1 : 0;
-const resourceInformationIndex = hasPreview ? 2 : (hasSchema ? 1 : 0);
-
-const { asc, getIdFromIndex, isSelected, selectIndex } = useTabs(tabsRef, tabListRef, resourceInformationIndex);
-
-function expand() {
-  if(expanded.value) {
-    globalThis._paq?.push(['trackEvent', 'Close resource', props.resource.id]);
+  if (open.value) {
+    trackEvent(['Open resource', props.resource.id]);
   } else {
-    globalThis._paq?.push(['trackEvent', 'Open resource', props.resource.id]);
+    trackEvent(['Close resource', props.resource.id]);
   }
-  expandAccordion();
-}
+};
 
-function registerEvent(tab: ComputedRef<string> | string) {
-  const tabName = unref(tab);
-  globalThis._paq?.push(['trackEvent', 'View resource tab', props.resource.id, tab]);
-  if(tabName === getIdFromIndex(resourcePreviewIndex)) {
-    globalThis._paq?.push(['trackEvent', 'Show preview', props.resource.id]);
-  } else if (tabName === getIdFromIndex(resourceStructureIndex)) {
-    globalThis._paq?.push(['trackEvent', 'Show data structure', props.resource.id]);
+const tabsOptions = computed(() => {
+  let options = [];
+
+  if (hasPreview.value) {
+    options.push({ key: "data", label: t("Data") })
+  }
+
+  if (props.resource.description) {
+    options.push({ key: "description", label: t("Description") })
+  }
+
+  if (hasPreview.value) {
+    options.push({ key: "data-structure", label: t("Data structure") })
+  }
+
+  options.push({ key: "metadata", label: t("Metadata") });
+  options.push({ key: "downloads", label: t("Downloads") });
+
+  return options;
+});
+const switchTab = (index: number) => {
+  const option = tabsOptions.value[index];
+  trackEvent(['View resource tab', props.resource.id, option.label])
+
+  if (option.key === 'data') {
+    trackEvent(['Show preview', props.resource.id])
+  }
+  if (option.key === 'data-structure') {
+    trackEvent(['Show data structure', props.resource.id])
   }
 }
 
-function getTabPanelClass(index: number) {
-  return {'fr-tabs__panel--selected': isSelected(index) };
-};
 
-function getTabPanelTabIndex(index: number) {
-  return isSelected(index) ? 0 : -1;
-};
+const communityResource = computed<CommunityResource | null>(() => {
+  if (! props.isCommunityResource) return null
+  return props.resource as CommunityResource
+})
+const owner = useOwnerName(communityResource.value);
 
-function getTabIdFromIndex(index: number) {
-  return getIdFromIndex(index) + "-tab";
-};
-
-function selectTab(index: number) {
-  selectIndex(index);
-  registerEvent(getIdFromIndex(index));
-}
-
-const availabilityChecked: boolean = hasExtras && 'check:available' in props.resource.extras;
 const lastUpdate = props.resource.last_modified;
+const availabilityChecked = props.resource.extras && 'check:available' in props.resource.extras;
 const unavailable = availabilityChecked && props.resource.extras['check:available'] === false;
-const { authorizeValidation, documentationUrl, loading, validationUrl, schemaReport} = useSchema(props.resource);
-const hasSchemaErrors = !!schemaReport.value.size;
+
 const resourceExternalUrl = computed(() => `${window.location.origin}${window.location.pathname}#/${props.isCommunityResource ? 'community-resources' : 'resources'}/${props.resource.id}`);
 
 const resourceContentId = 'resource-' + props.resource.id;
 const resourceHeaderId = 'resource-' + props.resource.id + '-header';
 const resourceTitleId = 'resource-' + props.resource.id + '-title';
-const resourceCopyId = 'resource-' + props.resource.id + '-copy';
-
-const { expanded, expand: expandAccordion } = useAccordion(resourceContentId);
-
-onMounted(() => {
-  if(props.expandedOnMount) {
-    expand();
-  }
-  if(copyRef.value && config.show_copy_resource_permalink) {
-    new Clipboard(copyRef.value);
-  }
-});
-
-const values = { true: '100%', false: '-100%' };
-// @ts-ignore this will be fine
-const translateValueFrom = computed(() => values[String(asc.value)]);
-// @ts-ignore this will be fine
-const translateValueTo = computed(() => values[String(!asc.value)]);
 </script>
-
 <style scoped>
-.slide-fade-enter-active,
-.slide-fade-leave-active {
-  transition: all 0.3s;
+.fr-link--no-after::after {
+  display: none !important;
 }
 
-.slide-fade-enter-from {
-  transform: translateX(v-bind(translateValueFrom));
-  opacity: 0;
+header:hover {
+  background-color: #f6f6f6;
 }
-.slide-fade-leave-to {
-  transform: translateX(v-bind(translateValueTo));
-  opacity: 0;
+
+/**
+If we do not put z-index, the header is fully clickable except for the DSFR icons (bad because one of the icons is the chevron up/down). It may be due to the usage of ::before to add the icon in the markup or the `mask-image`. We need to put a `z-2` on all elements that we want to be clickable over the header.
+*/
+.z-1 {
+  z-index: 1;
 }
+.z-2 {
+  z-index: 2;
+}
+
+.z-3 {
+  z-index: 3;
+}
+
+article {
+  container-type: inline-size;
+}
+
+@container (max-width: 600px) {
+  article header.flex {
+    flex-direction: column;
+    align-items: start;
+    justify-content: start;
+  }
+  article header .buttons {
+    margin-top: 1.25rem;
+    margin-left: auto !important;
+  }
+/*
+  If we want to put subheaders info in column on mobile…
+  article header .subheaders-infos {
+    display: flex;
+    flex-direction: column
+  }
+  article header .subheaders-infos .hidden.show-on-small {
+    display: inline !important;
+  }
+  article header .dash-after::after {
+    content: ''
+  } */
+
+  /* article .fr-pl-4v fr-pr-4v {
+    padding: 0.75rem !important;
+  } */
+}
+
 </style>