+++ udata_front/theme/gouvfr/datagouv-components/CHANGELOG.md
@@ -2,6 +2,14 @@
 
 ## Current (in progress)
 
+- Nothing yet
+
+## 1.3.1 (2024-09-12)
+
+- Fix toggletip and use summarize for metrics [#512](https://github.com/datagouv/udata-front/pull/512)
+
+## 1.3.0 (2024-09-10)
+
 - Use resource downloads for download icon on dataset card [#484](https://github.com/datagouv/udata-front/pull/484)
 - Reduce organization type icon [#483](https://github.com/datagouv/udata-front/pull/483)
 - add Tabs component [#491](https://github.com/datagouv/udata-front/pull/491)
@@ -10,6 +18,9 @@
 - add url to reuse type [#494](https://github.com/datagouv/udata-front/pull/494)
 - fix reuse card date shown [#496](https://github.com/datagouv/udata-front/pull/496)
 - add vue-clamp to cards title and description [#497](https://github.com/datagouv/udata-front/pull/497)
+- fix dsfr CSS for badge imported into components [#500](https://github.com/datagouv/udata-front/pull/500)
+- update DSFR to v1.11 [#466](https://github.com/datagouv/udata-front/pull/466) [#506](https://github.com/datagouv/udata-front/pull/506)
+- new resource accordion [#490](https://github.com/datagouv/udata-front/pull/490)
 
 ## 1.2.0 (2024-08-21)